requirements.txt
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.15
PyYAML==5.3.1
Flask==1.1.1
boto3==1.13.8
Pandas==1.0.3
requests==2.23.0
PyMySQL==0.9.3
scikit-learn==0.23.1
pytest==5.4.2